simplified databases, there were to many staging points for data and following everything was getting confusing, shooting for mvp
