+++ models/parent.js
@@ -1,0 +1,19 @@
+module.exports = function (sequilize, DataTypes) {
+
+    var user = sequilize.define("user", {
+        firstName: {
+            type: DataTypes.STRING,
+            allowNull: false
+        },
+        lastName: {
+            type: DataTypes.STRING,
+            allowNull: false
+        },
+        userId: {
+            type: DataTypes.STRING,
+            allowNull: false
+        }
+    });
+    
+    return user;
+};
